big refactor, changed user and deck cards design

+++ src/pages/single-deck/delete_dialog/useDeleteDeckDialogVM.ts
@@ -1,50 +1,47 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useClient } from "../../../client/useClient";
 import { useNavigate } from "react-router-dom";
-import { getUserRoute } from "../../../components/utils/getUserRoute";
+import { getUserRoute } from "../../../utils/getUserRoute";
+import { IDeleteDeckDialog } from "./DeleteDeckDialog";
 
-export const useDeleteDeckDialogVM = () => {
+export const useDeleteDeckDialogVM = ({ deckId, deckImageId }: IDeleteDeckDialog) => {
     const navigate = useNavigate();
     const { getInstance } = useClient();
     const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
-    const [deckId, setDeckId] = useState<string>();
-
-    const loadDefaultValues = async () => {
-        const currentDeckId = new URL(location.href).searchParams.get('id')
-        if (currentDeckId) {
-            setDeckId(currentDeckId);
-        }
-    }
 
     const handleRemoveDeck = async () => {
-        const supabase = getInstance();
+        if (deckId) {
+            const supabase = getInstance();
 
-        const { removedLinks, removedLinksError } = await supabase.from('card_in_deck').delete().eq('deck_id', deckId);
-        console.log("Links Removal: ", removedLinks, removedLinksError);
+            const { removedLinks, removedLinksError } = await supabase.from('card_in_deck').delete().eq('deck_id', deckId);
+            console.log("Links Removal: ", removedLinks, removedLinksError);
 
-        const { removedDeck, removedDeckError } = await supabase.from('deck').delete().eq('id', deckId);
-        console.log("Deck Removal: ", removedDeck, removedDeckError);
+            const { removedDeck, removedDeckError } = await supabase.from('deck').delete().eq('id', deckId);
+            console.log("Deck Removal: ", removedDeck, removedDeckError);
 
-        const removed = await handleRemoveDeckImage();
-        console.log("Deck Image Removal: ", removed);
+            const removed = await handleRemoveDeckImage();
+            console.log("Deck Image Removal: ", removed);
+        } else console.log("COULD NOT DELETE DECK, undefined id")
     }
 
     const handleRemoveDeckImage = async () => {
-        const supabase = getInstance();
-        const { data: deletionData, error: deletionError } =
-            await supabase.storage.from('DeckImages').remove([`${deckId}.jpeg`]);
+        if (deckImageId !== undefined && deckImageId !== null) {
+            const supabase = getInstance();
+            const { data: deletionData, error: deletionError } =
+                await supabase.storage.from('DeckImages').remove([deckImageId]);
 
-        console.log(deletionData)
-        deletionError && console.log("File upload error: ", deletionError);
-        return !deletionError;
+            console.log(deletionData)
+            deletionError && console.log("File deletion error: ", deletionError);
+            return !deletionError;
+        }
+
+        return "Not removed / obj = " + deckImageId;
     }
 
     const onDeleteSubmit = () => {
         handleRemoveDeck()
             .then(() => navigate(getUserRoute()));
     }
-
-    useEffect(() => { loadDefaultValues(); }, []);
 
     return {
         deleteDialogOpen,
